+++ syncrclone/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '20210928.0.BETA'
+__version__ = '20210930.0.BETA'
 LASTRCLONE = '1.56.1' # This is the last version I tested with. Does *NOT* mean it won't work further.
 
 import time